+++ src/vector/iterator.rs
@@ -1,95 +1,33 @@
+//! iterating over `Vector`s
+
 use crate::vector::Vector;
 use std::iter::FusedIterator;
-
-/// an implementation of `VectorIndexIterator`
-/// that starts at index 0 and moves forward until the last element
-#[derive(Debug)]
-pub struct ForwardVectorIndexIterator<const ROWS: usize>(Option<usize>);
-
-impl<const ROWS: usize> Iterator for ForwardVectorIndexIterator<ROWS> {
-    type Item = usize;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let index = self.0;
-        self.0 = match self.0 {
-            Some(i) if i < ROWS - 1 => Some(i + 1),
-            Some(_) => None, // fuse to None after returning ROWS - 1
-            None => None,    // once None is returned, always return None
-        };
-        index
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (ROWS, Some(ROWS))
-    }
-}
-
-impl<const ROWS: usize> Default for ForwardVectorIndexIterator<ROWS> {
-    fn default() -> Self {
-        Self(Some(0))
-    }
-}
-
-impl<const ROWS: usize> ExactSizeIterator for ForwardVectorIndexIterator<ROWS> {}
-
-impl<const ROWS: usize> FusedIterator for ForwardVectorIndexIterator<ROWS> {}
-
-/// an implementation of `VectorIndexIterator`
-/// that starts at the last index and moves backwards until index 0
-#[derive(Debug)]
-pub struct BackwardVectorIndexIterator<const ROWS: usize>(Option<usize>);
-
-impl<const ROWS: usize> Iterator for BackwardVectorIndexIterator<ROWS> {
-    type Item = usize;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let index = self.0;
-        self.0 = match self.0 {
-            Some(i) if i > 0 => Some(i - 1),
-            Some(_) => None, // fuse to None after returning 0
-            None => None,    // once None is returned, always return None
-        };
-        index
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (ROWS, Some(ROWS))
-    }
-}
-
-impl<const ROWS: usize> Default for BackwardVectorIndexIterator<ROWS> {
-    fn default() -> Self {
-        Self(Some(ROWS))
-    }
-}
-
-impl<const ROWS: usize> ExactSizeIterator for BackwardVectorIndexIterator<ROWS> {}
-
-impl<const ROWS: usize> FusedIterator for BackwardVectorIndexIterator<ROWS> {}
+use std::ops::Range;
 
 /// an iterator for consuming the elements of a `Vector`
 #[derive(Debug)]
-pub struct VectorIterator<T, I, const ROWS: usize>
+pub struct VectorIterator<T, const ROWS: usize, I>
 where
     T: Copy,
-    I: Iterator<Item = usize> + FusedIterator,
+    I: Iterator<Item = usize>,
 {
-    vector: Vector<T, ROWS>,
-    row_iterator: I,
+    pub(crate) vector: Vector<T, ROWS>,
+    pub(crate) row_iterator: I,
 }
 
-impl<T, I, const ROWS: usize> Iterator for VectorIterator<T, I, ROWS>
+impl<T, const ROWS: usize, I> Iterator for VectorIterator<T, ROWS, I>
 where
     T: Copy,
-    I: Iterator<Item = usize> + FusedIterator,
+    I: Iterator<Item = usize>,
 {
     type Item = T;
 
     fn next(&mut self) -> Option<Self::Item> {
         let index = match self.row_iterator.next() {
             Some(i) => i,
-            _ => return None,
+            None => return None,
         };
+        // we assume the index given is in bounds (if not, this panics)
         Some(self.vector[index])
     }
 
@@ -98,17 +36,32 @@ where
     }
 }
 
-impl<T, I, const ROWS: usize> ExactSizeIterator for VectorIterator<T, I, ROWS>
+impl<T, const ROWS: usize, I> DoubleEndedIterator for VectorIterator<T, ROWS, I>
 where
     T: Copy,
-    I: Iterator<Item = usize> + FusedIterator + ExactSizeIterator,
+    I: Iterator<Item = usize> + DoubleEndedIterator,
 {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        let index = match self.row_iterator.next_back() {
+            Some(i) => i,
+            None => return None,
+        };
+        // we assume the index given is in bounds (if not, this panics)
+        Some(self.vector[index])
+    }
 }
 
-impl<T, I, const ROWS: usize> FusedIterator for VectorIterator<T, I, ROWS>
+impl<T, const ROWS: usize, I> FusedIterator for VectorIterator<T, ROWS, I>
 where
     T: Copy,
     I: Iterator<Item = usize> + FusedIterator,
+{
+}
+
+impl<T, const ROWS: usize, I> ExactSizeIterator for VectorIterator<T, ROWS, I>
+where
+    T: Copy,
+    I: Iterator<Item = usize> + ExactSizeIterator,
 {
 }
 
@@ -117,31 +70,31 @@ where
     T: Copy,
 {
     type Item = T;
-    type IntoIter = VectorIterator<T, ForwardVectorIndexIterator<ROWS>, ROWS>;
+    type IntoIter = VectorIterator<T, ROWS, Range<usize>>;
 
     fn into_iter(self) -> Self::IntoIter {
         Self::IntoIter {
             vector: self,
-            row_iterator: Default::default(),
+            row_iterator: 0..ROWS,
         }
     }
 }
 
 /// an iterator for reading (by reference) the elements of a `Vector`
 #[derive(Debug)]
-pub struct VectorRefIterator<'a, T, I, const ROWS: usize>
+pub struct VectorReferenceIterator<'a, T, const ROWS: usize, I>
 where
     T: Copy,
-    I: Iterator<Item = usize> + FusedIterator,
+    I: Iterator<Item = usize>,
 {
-    vector: &'a Vector<T, ROWS>,
-    row_iterator: I,
+    pub(crate) vector: &'a Vector<T, ROWS>,
+    pub(crate) row_iterator: I,
 }
 
-impl<'a, T, I, const ROWS: usize> Iterator for VectorRefIterator<'a, T, I, ROWS>
+impl<'a, T, const ROWS: usize, I> Iterator for VectorReferenceIterator<'a, T, ROWS, I>
 where
     T: Copy,
-    I: Iterator<Item = usize> + FusedIterator,
+    I: Iterator<Item = usize>,
 {
     type Item = &'a T;
 
@@ -150,6 +103,7 @@ where
             Some(i) => i,
             _ => return None,
         };
+        // we assume the index given is in bounds (if not, this panics)
         Some(&self.vector[index])
     }
 
@@ -158,17 +112,32 @@ where
     }
 }
 
-impl<T, I, const ROWS: usize> ExactSizeIterator for VectorRefIterator<'_, T, I, ROWS>
+impl<T, const ROWS: usize, I> DoubleEndedIterator for VectorReferenceIterator<'_, T, ROWS, I>
 where
     T: Copy,
-    I: Iterator<Item = usize> + FusedIterator + ExactSizeIterator,
+    I: Iterator<Item = usize> + DoubleEndedIterator,
 {
+    fn next_back(&mut self) -> Option<Self::Item> {
+        let index = match self.row_iterator.next_back() {
+            Some(i) => i,
+            None => return None,
+        };
+        // we assume the index given is in bounds (if not, this panics)
+        Some(&self.vector[index])
+    }
 }
 
-impl<T, I, const ROWS: usize> FusedIterator for VectorRefIterator<'_, T, I, ROWS>
+impl<T, I, const ROWS: usize> FusedIterator for VectorReferenceIterator<'_, T, ROWS, I>
 where
     T: Copy,
     I: Iterator<Item = usize> + FusedIterator,
+{
+}
+
+impl<T, I, const ROWS: usize> ExactSizeIterator for VectorReferenceIterator<'_, T, ROWS, I>
+where
+    T: Copy,
+    I: Iterator<Item = usize> + ExactSizeIterator,
 {
 }
 
@@ -177,12 +146,12 @@ where
     T: Copy,
 {
     type Item = &'a T;
-    type IntoIter = VectorRefIterator<'a, T, ForwardVectorIndexIterator<ROWS>, ROWS>;
+    type IntoIter = VectorReferenceIterator<'a, T, ROWS, Range<usize>>;
 
     fn into_iter(self) -> Self::IntoIter {
         Self::IntoIter {
             vector: self,
-            row_iterator: Default::default(),
+            row_iterator: 0..ROWS,
         }
     }
 }